adds runnable real test
